requirements.txt
feedparser==6.0.11
pymisp==2.5.*
requests
python-dateutil
beautifulsoup4==4.13.5
iocextract==1.16.1
pymispwarninglists
openai
python-dotenv
google-genai
pytest
ruff